+++ lib/OpenCloud/ObjectStore/DataObject.php
@@ -161,11 +161,6 @@ class DataObject extends ObjectStore
 
         }
 
-        // flag missing Content-Type
-        if (empty($this->content_type)) {
-            $this->content_type = 'application/octet-stream';
-        }
-
         // Only allow supported archive types
         // http://docs.rackspace.com/files/api/v1/cf-devguide/content/Extract_Archive-d1e2338.html
         $extractArchiveUrlArg = '';
@@ -186,7 +181,10 @@ class DataObject extends ObjectStore
             $headers['ETag'] = $this->etag;
         }
 
-        $headers['Content-Type'] = $this->content_type;
+		// Content-Type is no longer required; if not specified, it will
+		// attempt to guess based on the file extension.
+		if (isset($this->content_type))
+        	$headers['Content-Type'] = $this->content_type;
         $headers['Content-Length'] = $this->content_length;
 
         // copy any extra headers